+++ database/migrations/2024_12_08_000001_add_phone_and_data_to_users_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
 	{
 		Schema::table('users', function (Blueprint $table)
 		{
-			$table->integer('phone')->nullable()->after('email');
+			$table->bigInteger('phone')->nullable()->after('email');
 			$table->json('data')->nullable()->after('phone');
 		});
 	}